Updates conf.py to the latest version
The version of conf.py before this commit was manually upgraded when I upgraded
Nikola to version 6. This manual upgrade was incomplete, which lead to
incorrect RSS links. To make sure this will not happen again, I created a
default conf.py and set the values that needed to be set.